make edit from code review
