+++ src/lib/drizzle.ts
@@ -1,5 +1,4 @@
 import { drizzle } from "drizzle-orm/postgres-js";
-import { sql } from "drizzle-orm";
 import postgres from "postgres";
 import * as schema from "./schema";
 
@@ -9,36 +8,3 @@ if (!process.env.DATABASE_URL) {
 
 export const connection = postgres(process.env.DATABASE_URL);
 export const db = drizzle(connection, { schema });
-
-// player queries
-export type NewPlayer = typeof schema.player.$inferInsert;
-export const insertPlayer = async (newPlayer: NewPlayer) => {
-  return db.insert(schema.player).values(newPlayer);
-};
-
-// game queries
-export type NewGame = typeof schema.game.$inferInsert;
-export const insertGame = async (newGame: NewGame) => {
-  return db.insert(schema.game).values(newGame);
-};
-export const checkDuplicatedGame = async (gameCode: string) => {
-  return db
-    .select({ gameCodeCount: sql<number>`count(*)`.mapWith(Number) })
-    .from(schema.game)
-    .where(
-      sql`${schema.game.code} = ${gameCode} AND (${
-        schema.game.status
-      } = ${"in progress"} OR ${schema.game.status} = ${"setting up"})`
-    );
-};
-
-export const getGame = async (gameCode: string) => {
-  return db
-    .select()
-    .from(schema.game)
-    .where(
-      sql`${schema.game.code} = ${gameCode} AND ${
-        schema.game.status
-      } = ${"setting up"}`
-    );
-};
